Moved makedb -> manage.py .. structured
Now manage.py can dump the currentl SQL structure, which is useful when
making changes to existing databases where we want to keep the data. In
the future we should definitely refactor this a bit to include a
config.py which abstracts the sqlite:/// stuff so we can rely on
different backends, etc without hardcoding sqlite/postgres stuff all
over.

Uses the click framework, which is why requirements.txt is updated.